the changes were created after checking Gleb(1)

+++ src/test/java/hw1/tests/MultiplyTest.java
@@ -6,10 +6,11 @@ import org.testng.annotations.Test;
 import static org.testng.Assert.assertEquals;
 
 public class MultiplyTest extends AbstractTest {
+
     @Test(dataProvider = "Multiplication Data Provider",
             dataProviderClass = MultiplicationDP.class)
-    public void multiplyTest(int a, int b, int expectedResult) {
-        int actualResult = Math.toIntExact(calculator.mult(a, b));
+    public void multiplyTest(long a, long b, long expectedResult) {
+        long actualResult = calculator.mult(a, b);
         assertEquals(actualResult, expectedResult);
     }
 }
